test: 쿠폰 발행 관련 문자열 레디스 기능 구현 및 테스트

+++ src/test/java/com/moabam/api/infrastructure/redis/StringRedisRepositoryTest.java
@@ -30,29 +30,21 @@ class StringRedisRepositoryTest {
 
 	@AfterEach
 	void setDown() {
-		stringRedisRepository.delete(key);
+		if (stringRedisRepository.hasKey(key)) {
+			stringRedisRepository.delete(key);
+		}
 	}
 
 	@DisplayName("레디스에 문자열 데이터가 성공적으로 저장된다. - Void")
 	@Test
-	void stringRedisRepository_save() {
+	void save_success() {
 		// Then
 		assertThat(stringRedisRepository.get(key)).isEqualTo(value);
 	}
 
-	@DisplayName("레디스의 특정 데이터가 성공적으로 삭제된다. - Void")
-	@Test
-	void stringRedisRepository_delete() {
-		// When
-		stringRedisRepository.delete(key);
-
-		// Then
-		assertThat(stringRedisRepository.hasKey(key)).isFalse();
-	}
-
 	@DisplayName("레디스의 특정 데이터가 성공적으로 조회된다. - String(Value)")
 	@Test
-	void stringRedisRepository_get() {
+	void get_success() {
 		// When
 		String actual = stringRedisRepository.get(key);
 
@@ -62,8 +54,28 @@ class StringRedisRepositoryTest {
 
 	@DisplayName("레디스의 특정 데이터 존재 여부를 성공적으로 체크한다. - Boolean")
 	@Test
-	void stringRedisRepository_hasKey() {
+	void hasKey_success() {
 		// When & Then
 		assertThat(stringRedisRepository.hasKey("not found key")).isFalse();
+	}
+
+	@DisplayName("레디스의 특정 데이터가 성공적으로 삭제된다. - Void")
+	@Test
+	void delete_success() {
+		// When
+		stringRedisRepository.delete(key);
+
+		// Then
+		assertThat(stringRedisRepository.hasKey(key)).isFalse();
+	}
+
+	@DisplayName("레디스의 특정 데이터의 값이 1 증가한다.")
+	@Test
+	void increment_success() {
+		// When
+		Long actual = stringRedisRepository.increment(key + "_INCR");
+
+		// Then
+		assertThat(actual).isEqualTo(1L);
 	}
 }
